Completely overhauled QuestionRoot and how question data is managed between scenes. It was very messy and needed to be redone so things would work more smoothly and work better together. This change also allowed backend and frontend code to stay more separated.
QuizManager class was added with the capability of storing and managing the flow of questions to be displayed to the screen. The .next() and .prev() methods manage the flow of questions
and what is displayed to the QuestionRoot screen. The .loadQuiz() method is the main method that establishes a quiz to be managed. I also made this class modular so it can be used with any QuestionRoot instance and
any class can instantiate and load data into the manager.

To keep track of questions for the program's life cycle, I created a static instance of the manager in 'Main' where you can access it publicly from any other class.